code/DeepDG/requirements.txt
Please refer to https://hub.docker.com/r/jindongwang/docker.

CUDA: 10.1
Pytorch: 1.7
Torchvision: 0.8.1
